+++ src/alcotest-engine/cli.ml
@@ -76,7 +76,7 @@ module Make (P : Platform.MAKER) (M : Monad.S) :
       alcotest_columns;
     ]
 
-  let set_color =
+  let set_color stdout stderr =
     let env = Cmd.Env.info "ALCOTEST_COLOR" in
     let+ color_flag =
       let enum = [ ("auto", `Auto); ("always", `Ansi_tty); ("never", `None) ] in
@@ -102,17 +102,20 @@ module Make (P : Platform.MAKER) (M : Monad.S) :
             Some `Ansi_tty
           with Not_found -> None)
     in
-    P.setup_std_outputs ?style_renderer ()
+    P.setup_std_outputs ?style_renderer stdout stderr
 
   let default_cmd config args library_name tests =
     let and_exit = Config.User.and_exit config
     and record_backtrace = Config.User.record_backtrace config
-    and ci = Config.User.ci config in
+    and ci = Config.User.ci config
+    and stdout = Config.User.stdout config
+    and stderr = Config.User.stderr config in
     let exec_name = Filename.basename Sys.argv.(0) in
     let doc = "Run all the tests." in
     let term =
-      let+ () = set_color
-      and+ cli_config = Config.User.term ~and_exit ~record_backtrace ~ci
+      let+ () = set_color stdout stderr
+      and+ cli_config =
+        Config.User.term ~stdout ~stderr ~and_exit ~record_backtrace ~ci
       and+ args = args in
       let config = Config.User.(cli_config || config) in
       run_with_args' config library_name args tests
@@ -120,31 +123,36 @@ module Make (P : Platform.MAKER) (M : Monad.S) :
     (term, Cmd.info exec_name ~doc ~envs)
 
   let test_cmd config args library_name tests =
-    let ci = Config.User.ci config in
+    let ci = Config.User.ci config
+    and stdout = Config.User.stdout config
+    and stderr = Config.User.stderr config in
     let doc = "Run a subset of the tests." in
     let term =
-      let+ () = set_color
+      let+ () = set_color stdout stderr
       and+ cli_config =
-        Config.User.term ~and_exit:true ~record_backtrace:true ~ci
+        Config.User.term ~stdout ~stderr ~and_exit:true ~record_backtrace:true
+          ~ci
       and+ args = args in
       let config = Config.User.(cli_config || config) in
       run_with_args' config library_name args tests
     in
     (term, Cmd.info "test" ~doc ~envs)
 
-  let list_cmd tests =
+  let list_cmd ~stdout ~stderr tests =
     let doc = "List all available tests." in
-    ( (let+ () = set_color in
+    ( (let+ () = set_color stdout stderr in
        list_tests tests),
       Cmd.info "list" ~doc )
 
   let run_with_args' (type a) ~argv config name (args : a Term.t)
       (tl : a test list) =
     let ( >>= ) = M.bind in
+    let stdout = Config.User.stdout config
+    and stderr = Config.User.stderr config in
     let choices =
       List.map
         (fun (term, info) -> Cmd.v info term)
-        [ list_cmd tl; test_cmd config args name tl ]
+        [ list_cmd ~stdout ~stderr tl; test_cmd config args name tl ]
     in
     let and_exit = Config.User.and_exit config in
     let exit_or_return exit_code =
@@ -162,11 +170,12 @@ module Make (P : Platform.MAKER) (M : Monad.S) :
     | Error (`Parse | `Term) -> exit_or_return Cmd.Exit.cli_error
     | Error `Exn -> exit Cmd.Exit.internal_error
 
-  let run_with_args ?and_exit ?verbose ?compact ?tail_errors ?quick_only
-      ?show_errors ?json ?filter ?log_dir ?bail ?record_backtrace ?ci ?argv =
-    Config.User.kcreate (run_with_args' ~argv) ?and_exit ?verbose ?compact
-      ?tail_errors ?quick_only ?show_errors ?json ?filter ?log_dir ?bail
-      ?record_backtrace ?ci
+  let run_with_args ?stdout ?stderr ?and_exit ?verbose ?compact ?tail_errors
+      ?quick_only ?show_errors ?json ?filter ?log_dir ?bail ?record_backtrace
+      ?ci ?argv =
+    Config.User.kcreate (run_with_args' ~argv) ?stdout ?stderr ?and_exit
+      ?verbose ?compact ?tail_errors ?quick_only ?show_errors ?json ?filter
+      ?log_dir ?bail ?record_backtrace ?ci
 
   let run =
     Config.User.kcreate (fun config ?argv name tl ->